requirements.txt
pymongo>=4.6.0
chromadb>=0.4.0
neo4j>=5.14.0
graphiti-core>=0.22.0
pydantic>=2.0.0
openai>=2.6.0
anthropic>=0.71.0
google-genai>=1.45.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-louvain>=0.16
networkx>=3.0
